Fix compiler issues
